Added messageservice

+++ LiveWriterPluginManager/ViewModel/PluginViewModel.cs
@@ -9,10 +9,13 @@ namespace LiveWriterPluginManager.ViewModel
     public class PluginViewModel : ViewModelBase
     {
         private readonly ILiveWriterService _liveWriterService;
-        public PluginViewModel(Plugin plugin, ILiveWriterService liveWriterService)
+        private readonly IMessageService _messageService;
+
+        public PluginViewModel(Plugin plugin, ILiveWriterService liveWriterService, IMessageService messageService)
         {
             Plugin = plugin;
             _liveWriterService = liveWriterService;
+            _messageService = messageService;
         }
 
         public Plugin Plugin { get; set; }
@@ -21,14 +24,19 @@ namespace LiveWriterPluginManager.ViewModel
         {
             get
             {
-                return new RelayCommand(() =>
+                return new RelayCommand(async () =>
                 {
                     if (AppHelper.IsLiveWriterRunning())
                     {
                         // TODO: Display an error message
                         return;
                     }
-                    _liveWriterService.DeletePlugin(Plugin);
+
+                    var removePlugin = await _messageService.ShowAsync("Are you sure?", "Are you sure you wish to remove this plugin?");
+                    if (removePlugin)
+                    {
+                        _liveWriterService.DeletePlugin(Plugin);
+                    }
                 });
             }
         }
